Update tbitfield.h: TELEM as uint

+++ include/tbitfield.h
@@ -12,7 +12,7 @@
 
 using namespace std;
 
-typedef unsigned long long TELEM;
+typedef unsigned int TELEM;
 
 class TBitField
 {
